+++ app/assets/javascripts/actions/actionTypes.js
@@ -1,0 +1,13 @@
+import keyMirror from 'keymirror';
+
+export default keyMirror({
+  REQUEST_PROJECT_BOARD: null,
+  RECEIVE_PROJECT_BOARD: null,
+  ERROR_REQUEST_PROJECT_BOARD: null,
+  RECEIVE_PROJECT: null,
+  RECEIVE_USERS: null,
+  RECEIVE_STORIES: null,
+  COLUMN_CHILLY_BIN: null,
+  COLUMN_BACKLOG: null,
+  COLUMN_IN_PROGRESS: null,
+});