+++ app/assets/javascripts/reducers/columns/index.js
@@ -1,0 +1,12 @@
+import { combineReducers } from 'redux';
+import backlog from './backlog';
+import chillyBin from './chillyBin';
+import done from './done';
+import inProgress from './inProgress';
+
+export default combineReducers({
+  backlog,
+  chillyBin,
+  done,
+  inProgress,
+});